refactor: relocate binary-shell helpers (logging + process lock + recovery dir) [#512]
Lifts the binary-shell helpers that were transiting through mokumo-api:

- services/api/src/logging.rs -> crates/kikan/src/logging.rs
  Tracing init (console + JSON file layers, --verbose/--quiet flag mapping).
  No domain language inside; cleanly fits I1. Adds tracing-subscriber +
  tracing-appender deps on kikan.

- resolve_recovery_dir, DRAIN_TIMEOUT_SECS, lock_file_path,
  read_lock_info, write_lock_info, format_lock_conflict_message,
  format_reset_db_conflict_message -> crates/mokumo-shop/src/startup.rs
  Placed alongside the other startup helpers so BDD tests can reach them
  once they migrate to mokumo-shop/tests in B3-bulk.

Binaries (mokumo-server, mokumo-desktop) repointed to the new homes.
services/api retains transitional `pub use` re-exports so BDD tests
continue compiling until B3-bulk migrates them.

Refs #512 (Stage 8, part B2b).
