+++ crates/kikan/Cargo.toml
@@ -20,6 +20,8 @@ tokio = { workspace = true }
 tokio-util = { workspace = true }
 kikan-types = { workspace = true }
 tracing = { workspace = true }
+tracing-appender = { workspace = true }
+tracing-subscriber = { workspace = true }
 parking_lot = { workspace = true }
 tower = { workspace = true }
 tower-http = { workspace = true }
