+++ crates/mokumo-shop/Cargo.toml
@@ -35,6 +35,7 @@ tokio = { workspace = true }
 tracing = { workspace = true }
 uuid = { workspace = true }
 rusqlite = { workspace = true }
+dirs = { workspace = true }
 tempfile = "3"
 tower-sessions-sqlx-store = { workspace = true }
 url = "2"
